Create second project for mypy-protobuf
This was a rudimentary way of having multiple mypy commands
per project (as needed here). It might be more elegant to have
multiple mypy_cmd within a project, but it seemed nontrivial to
achieve, and I was able to whip this mechanism up more quickly -
using multiple projects with one mypy_cmd per project.
Not the most proud of it, but it works.